Newest Version (01.06.22)

+++ Assets/Scripts/EnemyHUD.cs
@@ -18,13 +18,13 @@ public class EnemyHUD : MonoBehaviour
         health.value = attribute.currentHP;
     }
 
-    public void setMaxHP(int hp) {
+    public void setMaxHP(float hp) {
 
         health.maxValue = hp;
         health.value = hp;
     }
 
-    public void setHP(int hp) {
+    public void setHP(float hp) {
 
         health.value = hp;
     }
